+++ packages/ws-worker/test/events/step-complete.test.ts
@@ -1,29 +1,29 @@
 import test from 'ava';
-import handleRunStart from '../../src/events/run-complete';
+import handleStepStart from '../../src/events/step-complete';
 
 import { mockChannel } from '../../src/mock/sockets';
 import { createAttemptState } from '../../src/util';
-import { RUN_COMPLETE } from '../../src/events';
+import { STEP_COMPLETE } from '../../src/events';
 
 import type { ExecutionPlan } from '@openfn/runtime';
 
-test('clear the run id and active job on state', async (t) => {
+test('clear the step id and active job on state', async (t) => {
   const plan = { id: 'attempt-1' };
   const jobId = 'job-1';
 
   const state = createAttemptState(plan);
   state.activeJob = jobId;
-  state.activeRun = 'b';
+  state.activeStep = 'b';
 
   const channel = mockChannel({
-    [RUN_COMPLETE]: () => true,
+    [STEP_COMPLETE]: () => true,
   });
 
   const event = { state: { x: 10 } };
-  await handleRunStart({ channel, state }, event);
+  await handleStepStart({ channel, state }, event);
 
   t.falsy(state.activeJob);
-  t.falsy(state.activeRun);
+  t.falsy(state.activeStep);
 });
 
 test('setup input mappings on on state', async (t) => {
@@ -33,16 +33,16 @@ test('setup input mappings on on state', async (t) => {
 
   const state = createAttemptState(plan);
   state.activeJob = jobId;
-  state.activeRun = 'b';
+  state.activeStep = 'b';
 
   const channel = mockChannel({
-    [RUN_COMPLETE]: (evt) => {
+    [STEP_COMPLETE]: (evt) => {
       lightningEvent = evt;
     },
   });
 
   const engineEvent = { state: { x: 10 }, next: ['job-2'] };
-  await handleRunStart({ channel, state }, engineEvent);
+  await handleStepStart({ channel, state }, engineEvent);
 
   t.deepEqual(state.inputDataclips, {
     ['job-2']: lightningEvent.output_dataclip_id,
@@ -55,14 +55,14 @@ test('save the dataclip to state', async (t) => {
 
   const state = createAttemptState(plan);
   state.activeJob = jobId;
-  state.activeRun = 'b';
+  state.activeStep = 'b';
 
   const channel = mockChannel({
-    [RUN_COMPLETE]: () => true,
+    [STEP_COMPLETE]: () => true,
   });
 
   const event = { state: { x: 10 } };
-  await handleRunStart({ channel, state }, event);
+  await handleStepStart({ channel, state }, event);
 
   t.is(Object.keys(state.dataclips).length, 1);
   const [dataclip] = Object.values(state.dataclips);
@@ -75,16 +75,16 @@ test('write a reason to state', async (t) => {
 
   const state = createAttemptState(plan);
   state.activeJob = jobId;
-  state.activeRun = 'b';
+  state.activeStep = 'b';
 
   t.is(Object.keys(state.reasons).length, 0);
 
   const channel = mockChannel({
-    [RUN_COMPLETE]: () => true,
+    [STEP_COMPLETE]: () => true,
   });
 
   const event = { state: { x: 10 } };
-  await handleRunStart({ channel, state }, event);
+  await handleStepStart({ channel, state }, event);
 
   t.is(Object.keys(state.reasons).length, 1);
   t.deepEqual(state.reasons[jobId], {
@@ -100,17 +100,17 @@ test('generate an exit reason: success', async (t) => {
 
   const state = createAttemptState(plan);
   state.activeJob = jobId;
-  state.activeRun = 'b';
+  state.activeStep = 'b';
 
   let event;
 
   const channel = mockChannel({
-    [RUN_COMPLETE]: (e) => {
+    [STEP_COMPLETE]: (e) => {
       event = e;
     },
   });
 
-  await handleRunStart({ channel, state }, { state: { x: 10 } });
+  await handleStepStart({ channel, state }, { state: { x: 10 } });
 
   t.truthy(event);
   t.is(event.reason, 'success');
@@ -118,19 +118,19 @@ test('generate an exit reason: success', async (t) => {
   t.is(event.error_message, null);
 });
 
-test('send a run:complete event', async (t) => {
+test('send a step:complete event', async (t) => {
   const plan = { id: 'attempt-1' };
   const jobId = 'job-1';
   const result = { x: 10 };
 
   const state = createAttemptState(plan);
   state.activeJob = jobId;
-  state.activeRun = 'b';
+  state.activeStep = 'b';
 
   const channel = mockChannel({
-    [RUN_COMPLETE]: (evt) => {
+    [STEP_COMPLETE]: (evt) => {
       t.is(evt.job_id, jobId);
-      t.truthy(evt.run_id);
+      t.truthy(evt.step_id);
       t.truthy(evt.output_dataclip_id);
       t.is(evt.output_dataclip, JSON.stringify(result));
       t.deepEqual(evt.mem, event.mem);
@@ -146,5 +146,5 @@ test('send a run:complete event', async (t) => {
     duration: 61,
     threadId: 'abc',
   };
-  await handleRunStart({ channel, state }, event);
+  await handleStepStart({ channel, state }, event);
 });